ml-service/requirements.txt
# -----------------------------
# FastAPI + Web Stack
# -----------------------------
fastapi>=0.103
uvicorn[standard]>=0.23
python-multipart>=0.0.6
pydantic>=2.4

# -----------------------------
# NLP / Transformers
# -----------------------------
transformers>=4.38
sentence-transformers>=2.6.0
tokenizers>=0.15.0

# -----------------------------
# Embeddings, Vector Search, Similarity
# (Using FAISS CPU for Windows)
# -----------------------------
faiss-cpu>=1.8.0

# -----------------------------
# NLP Utilities
# -----------------------------
spacy>=3.7
nltk>=3.8
scikit-learn>=1.3
tqdm>=4.66

# -----------------------------
# Data Handling
# -----------------------------
numpy>=1.26
pandas>=2.2
pyyaml>=6.0

# -----------------------------
# Ingestion / Web Scraping
# -----------------------------
feedparser>=6.0
arxiv>=1.6
newspaper3k>=0.2.8
requests>=2.31

# -----------------------------
# Logging / Utilities
# -----------------------------
loguru>=0.7
python-dotenv>=1.0

# -----------------------------
# Optional (DB, if needed)
# -----------------------------
pymongo>=4.5      # remove if not using MongoDB
sqlalchemy>=2.0   # remove if not using SQL DB

# -----------------------------
# Development Tools (Optional)
# -----------------------------
black
ruff
isort
